streamlit-component/requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
snowflake-connector-python==3.6.0
python-dotenv==1.0.1 